comparison printing and timing

+++ concurrency/csc410/src/calclib/formula.rs
@@ -1,12 +1,12 @@
 #[allow(dead_code)]
-pub fn square(x: f32) -> f32 {
+pub fn square(x: f64) -> f64 {
   return x.powf(2.0);
 }
 #[allow(dead_code)]
-pub fn sin(x: f32) -> f32 {
+pub fn sin(x: f64) -> f64 {
   return x.sin();
 }
 #[allow(dead_code)]
-pub fn exp(x: f32) -> f32 {
+pub fn exp(x: f64) -> f64 {
   return x.exp();
 }
